+++ aziot/Cargo.toml
@@ -15,11 +15,11 @@ structopt = "0.3"
 toml = "0.5"
 url = "2"
 
-aziot-certd = { path = "../cert/aziot-certd" }
-aziot-identityd = { path = "../identity/aziot-identityd" }
-aziot-keyd = { path = "../key/aziot-keyd" }
+aziot-certd-config = { path = "../cert/aziot-certd-config" }
+aziot-identityd-config = { path = "../identity/aziot-identityd-config" }
+aziot-keyd-config = { path = "../key/aziot-keyd-config" }
 aziot-keys-common = { path = "../key/aziot-keys-common" }
-aziot-tpmd = { path = "../tpm/aziot-tpmd" }
+aziot-tpmd-config = { path = "../tpm/aziot-tpmd-config" }
 
 
 [dev-dependencies]
