+++ cert/aziot-cert-client-async/Cargo.toml
@@ -13,4 +13,4 @@ serde_json = "1"
 
 aziot-cert-common-http = { path = "../aziot-cert-common-http" }
 aziot-key-common = { path = "../../key/aziot-key-common" }
-http-common = { path = "../../http-common" }
+http-common = { path = "../../http-common", features = ["tokio02"] }
