+++ cert/aziot-certd-config/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "aziot-certd-config"
+version = "0.1.0"
+authors = ["Azure IoT Edge Devs"]
+edition = "2018"
+
+
+[dependencies]
+http-common = { path = "../../http-common" }
+serde = { version = "1", features = ["derive"] }
+url = { version = "2", features = ["serde"] }
+
+
+[dev-dependencies]
+toml = "0.5"
