+++ http-common/Cargo.toml
@@ -24,5 +24,4 @@ url = { version = "2", features = ["serde"] }
 serde_json = "1"
 
 [features]
-default = ["tokio02"]
 tokio02 = ["hyper", "tokio"]
