+++ identity/aziot-identityd/Cargo.toml
@@ -11,23 +11,20 @@ edition = "2018"
 
 [dependencies]
 async-trait = "0.1"
-base64 = "0.12"
 chrono = "0.4"
 futures-util = "0.3"
 http = "0.2"
 hyper = "0.13"
-hyper-openssl = "0.8"
 lazy_static = "1"
 libc = "0.2"
 log = "0.4"
 openssl = "0.10"
 percent-encoding = "2"
 regex = "1"
-serde = { version = "1", features = ["derive"] }
+serde = "1"
 serde_json = "1.0"
-tokio = { version = "0.2", features = ["time"] }
 toml = "0.5"
-url = { version = "2.0", features = ["serde"] }
+url = "2"
 
 aziot-cert-client-async = { path = "../../cert/aziot-cert-client-async" }
 aziot-cert-common-http = { path = "../../cert/aziot-cert-common-http" }
@@ -35,18 +32,17 @@ aziot-dps-client-async = { path = "../aziot-dps-client-async" }
 aziot-hub-client-async = { path = "../aziot-hub-client-async" }
 aziot-identity-common = { path = "../aziot-identity-common" }
 aziot-identity-common-http = { path = "../aziot-identity-common-http" }
+aziot-identityd-config = { path = "../aziot-identityd-config" }
 aziot-key-client = { path = "../../key/aziot-key-client" }
 aziot-key-client-async = { path = "../../key/aziot-key-client-async" }
 aziot-key-common = { path = "../../key/aziot-key-common" }
 aziot-key-common-http = { path = "../../key/aziot-key-common-http" }
 aziot-key-openssl-engine = { path = "../../key/aziot-key-openssl-engine" }
 aziot-tpm-client-async = { path = "../../tpm/aziot-tpm-client-async" }
-aziot-tpm-common = { path = "../../tpm/aziot-tpm-common" }
 aziot-tpm-common-http = { path = "../../tpm/aziot-tpm-common-http" }
 http-common = { path = "../../http-common" }
 openssl2 = { path = "../../openssl2" }
 
 
 [dev-dependencies]
-test-case = "1.0.0"
 tokio = { version = "0.2", features = ["macros"] }
