+++ identity/aziot-identityd/src/http/create_or_list_module_identity.rs
@@ -3,7 +3,7 @@
 pub(super) struct Route {
     api: std::sync::Arc<futures_util::lock::Mutex<crate::Api>>,
     id_type: Option<String>,
-    user: crate::auth::Credentials,
+    user: aziot_identityd_config::Credentials,
 }
 
 #[async_trait::async_trait]
@@ -37,7 +37,7 @@ impl http_common::server::Route for Route {
         Some(Route {
             api: service.api.clone(),
             id_type,
-            user: crate::auth::Uid(uid),
+            user: aziot_identityd_config::Uid(uid),
         })
     }
