+++ tpm/aziot-tpmd-config/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "aziot-tpmd-config"
+version = "0.1.0"
+authors = ["Azure IoT Edge Devs"]
+edition = "2018"
+
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+
+http-common = { path = "../../http-common" }
+
+
+[dev-dependencies]
+toml = "0.5"
