+++ tpm/aziot-tpmd/Cargo.toml
@@ -6,22 +6,16 @@ edition = "2018"
 
 [dependencies]
 async-trait = "0.1"
-backtrace = "0.3"
-base64 = "0.12"
 futures-util = "0.3"
 http = "0.2"
 hyper = "0.13"
-lazy_static = "1"
 log = "0.4"
-percent-encoding = "2"
-regex = "1"
-serde = { version = "1", features = ["derive"] }
+serde = "1"
 serde_json = "1"
-tokio = { version = "0.2", features = ["macros"] }
-toml = "0.5"
 url = "2"
 
 aziot-tpm = { path = "../aziot-tpm-rs" }
 aziot-tpm-common = { path = "../aziot-tpm-common" }
 aziot-tpm-common-http = { path = "../aziot-tpm-common-http" }
+aziot-tpmd-config = { path = "../aziot-tpmd-config" }
 http-common = { path = "../../http-common" }
